// js-packages/openk9-admin-new/src/graphql-generated.ts
export enum AnnotatorType {
  Aggregator = 'AGGREGATOR',
  Autocomplete = 'AUTOCOMPLETE',
  Autocorrect = 'AUTOCORRECT',
  Doctype = 'DOCTYPE',
  Keyword = 'KEYWORD',
  KeywordAutocomplete = 'KEYWORD_AUTOCOMPLETE',
  Ner = 'NER',
  NerAutocomplete = 'NER_AUTOCOMPLETE',
  Stopword = 'STOPWORD',
  Token = 'TOKEN'
}
// ...
export enum Fuzziness {
  Auto = 'AUTO',
  One = 'ONE',
  Two = 'TWO',
  Zero = 'ZERO'
}